Move __main__ content
Moved the content of __main__.py to the new module cli.py
